Add postgresql database
This adds a postgresql database for dynmap to use, as the data volume was a little slow this also adds a dedicated volume for postgres. The database should be far more optimised for this use case.

Additionally, going against terraform best practices, I've moved some service configuration to my service config module. Allowing config files to be updated and reloaded without a server recreation.