Move to t4g.medium for minecraft server
m6i.large seems to be the perfect choice for the server - it's able to handle everything it needs to. However, the server is staying online around 20 hours/day which means m6i.large incurrs a significant cost. Goign to experiment with t4g.medium and see if it's able to work. I don't expect the same performance, but it should hopefully be usable.